Module Dkml_install_runner.Os_utils

val string_to_norm_fpath : string -> Fpath.t

string_to_fpath str converts str into a Fpath.t. On Windows the str is normalized to a regular Windows file path (ex. backslashes).

val normalize_path : string -> string

normalize_path normalizes a path so on Windows it is a regular Windows path with backslashes.

type install_files_source =

| Opam_switch_prefix of Fpath.t

| Install_files_dir of Fpath.t

type install_files_type =

| Staging

| Static

type package_selector =

| Package

| Component

val absdir_install_files : ?package_selector:package_selector -> component_name:string -> install_files_type -> install_files_source -> Fpath.t

absdir_install_files ~component_name install_files_type install_files_source is the component_name component's static-files or staging-files directory for Staging or Static install_files_type, respectively