Module Runner.Cmdliner_runner

val help : [> `Pager ] as 'a -> string list -> string option -> [> `Error of bool * string | `Help of 'a * string option | `Ok of unit ]

val help_secs : [> `P of string | `S of string ] list

val setup_log : Fmt.style_renderer option -> Logs.level option -> unit

val setup_log_t : unit Cmdliner.Term.t

val create_context : string -> Dkml_install_register.Component_registry.t -> unit -> string -> string option -> bool -> Dkml_install_api.Context.t

val prefix_t : string Cmdliner.Term.t

val staging_files_opt_t : string option Cmdliner.Term.t

val opam_context_t : bool Cmdliner.Term.t

val ctx_t : string -> Dkml_install_register.Component_registry.t -> Dkml_install_api.Context.t Cmdliner.Term.t

ctx_t component creates a Term for component that sets up logging and any other global state, and defines the context record

val help_cmd : unit Cmdliner.Term.t * Cmdliner.Term.info