Module Runner.Os_utils

val string_to_norm_fpath : string -> Fpath.t

string_to_fpath str converts str into a Fpath.t. On Windows the str is normalized to a regular Windows file path (ex. backslashes).

val normalize_path : string -> string

normalize_path normalizes a path so on Windows it is a regular Windows path with backslashes.

val copy_dir : Fpath.t -> Fpath.t -> ( unit, string ) Stdlib.result

type install_files_source =

| Opam_context

| Install_files_dir of string

type install_files_type =

| Staging

| Static

val absdir_install_files : component_name:string -> install_files_type -> install_files_source -> string

absdir_install_files ~component_name install_files_type install_files_source is the component_name component's static-files or staging-files directory for Staging or Static install_files_type, respectively