Module Runner.Cmdliner_runner

val help : [> `Pager ] as 'a -> string list -> string option -> [> `Error of bool * string | `Help of 'a * string option | `Ok of unit ]

val help_secs : [> `P of string | `S of string ] list

val setup_log : Fmt.style_renderer option -> Logs.level option -> Runner__Cmdliner_common.log_config

val setup_log_t : Runner__Cmdliner_common.log_config Cmdliner.Term.t

val staging_files_source : opam_context:bool -> staging_files_opt:string option -> Path_eval.Global_context.staging_files_source

val create_context : string -> Dkml_install_register.Component_registry.t -> Runner__Cmdliner_common.log_config -> string -> string option -> bool -> Dkml_install_api.Context.t

val quote : string -> string

val err_no : string -> string -> string

val err_not_dir : string -> string

val raw_pp_str : Stdlib.Format.formatter -> string -> unit

val raw_dir : ( string -> [> `Error of string | `Ok of string ] ) * ( Stdlib.Format.formatter -> string -> unit )

val prefix_t : string Cmdliner.Term.t

val archive_dir_for_setup : Fpath.t

val staging_files_opt_t : string option Cmdliner.Term.t

val staging_files_for_setup_and_uninstaller_t : string Cmdliner.Term.t

staging_files_for_setup_and_uninstaller_t is the dkml-install-setup.exe Term for the staging files directory. It defaults to the sibling directory "staging".

val static_files_for_setup_and_uninstaller_t : string Cmdliner.Term.t

static_files_for_setup_and_uninstaller_t is the dkml-install-setup.exe Term for the static files directory. It defaults to the sibling directory "static".

val opam_context_t : bool Cmdliner.Term.t

val staging_files_source_for_setup_and_uninstaller_t : Path_eval.Global_context.staging_files_source Cmdliner.Term.t

val static_files_source_for_setup_and_uninstaller_t : Component_utils.static_files_source Cmdliner.Term.t

val ctx_t : string -> Dkml_install_register.Component_registry.t -> Dkml_install_api.Context.t Cmdliner.Term.t

ctx_t component creates a Term for component that sets up logging and any other global state, and defines the context record

val help_cmd : unit Cmdliner.Term.t * Cmdliner.Term.info